requirements.txt
# dev environment / analytics
jupyter
jupyterlab
matplotlib

# packages
web3
requests
omegaconf
python-dotenv
pandas